fix(http): improve error handling and type safety in HTTP transport layer

+++ exeris-kernel-community/src/main/java/eu/exeris/kernel/community/http/Http2RequestStreamState.java
@@ -77,7 +77,7 @@ final class Http2RequestStreamState implements AutoCloseable {
         int nextCapacity = bodyBuffer == null
                 ? INITIAL_BODY_BUFFER_BYTES
                 : (int) bodyBuffer.capacity();
-        while (nextCapacity < needed) {
+        while (nextCapacity < (int) needed) {
             nextCapacity = Math.max(nextCapacity * 2, (int) needed);
         }
